fix: calculation when slot is 22h00-00h30 should end on next day
Previously the GetLoadsheddingCalendar would result in very long loadshedding calendar events because a 22h00-00h30 slot would end up being rendered as 00h30-22h00 on the same day (on the day of 22h00). It will now instead roll the 00h30 over to the next day, by effectively adding 24 to the hour component (so making it 24h30).